+++ src/variants/no-bcd-signalling.js
@@ -1,9 +1,6 @@
 const fs = require("fs");
 const path = require("path");
 
-// const httpServer = require("http-server");
-// const minimalcss = require("minimalcss");
-// const puppeteer = require("puppeteer");
 const cheerio = require("cheerio");
 const { gzip } = require("node-gzip");
 
@@ -18,15 +15,17 @@ async function main(folder, options, logger) {
     const $element = $(element);
     let uri = $element.attr("src");
     if (uri && uri.includes("bcd-signal")) {
-      console.log(uri);
+      logger.info(`Removing BCD Signalling URI ${uri}`);
       $element.remove();
     }
   });
 
   let finalHtml = $.html();
   fs.writeFileSync(htmlFile, finalHtml);
-  const compressed = await gzip(finalHtml);
-  fs.writeFileSync(htmlFile + ".gz", compressed);
+  if (fs.existsSync(htmlFile + ".gz")) {
+    const compressed = await gzip(finalHtml);
+    fs.writeFileSync(htmlFile + ".gz", compressed);
+  }
 }
 
 module.exports = {
